feat: render react component w/ no props
