--- apps/angular/react-in-angular/src/app/react/post.component.ts
@@ -1,19 +1,0 @@
-import { Component, EventEmitter, input, Output } from '@angular/core';
-
-type Post = { title: string; description: string; pictureLink: string };
-
-// this is the wrapper component ?
-
-@Component({
-  standalone: true,
-  selector: 'app-post',
-  template: `
-    <div></div>
-  `,
-  styles: [''],
-})
-export class PostComponent {
-  post = input<Post | undefined>(undefined);
-  isSelected = input<boolean>(false);
-  @Output() selectPost = new EventEmitter<void>();
-}
