+++ apps/angular/react-in-angular/tsconfig.editor.json
@@ -1,6 +1,6 @@
 {
   "extends": "./tsconfig.json",
-  "include": ["src/**/*.ts"],
+  "include": ["src/**/*.ts", "src/app/react/post.component.tsx"],
   "compilerOptions": {
     "types": ["jest", "node"]
   }
