+++ package.json
@@ -35,6 +35,8 @@
     "@rx-angular/template": "^17.0.0",
     "@swc/helpers": "0.5.3",
     "@tanstack/angular-query-experimental": "^5.12.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "rxjs": "7.8.1",
     "tailwindcss": "^3.3.5",
     "tslib": "^2.3.0",
@@ -74,6 +76,8 @@
     "@testing-library/user-event": "^14.5.1",
     "@types/jest": "29.4.0",
     "@types/node": "18.16.9",
+    "@types/react": "^18.2.55",
+    "@types/react-dom": "^18.2.19",
     "@typescript-eslint/eslint-plugin": "6.11.0",
     "@typescript-eslint/parser": "6.11.0",
     "all-contributors-cli": "^6.26.1",
